requirements.txt
beautifulsoup4
chromadb
google-search-results
langchain
openai
pdfrw
pymupdf
selenium
selenium-requests
unstructured[local-inference]
django
djangorestframework
